# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Sentinel for Caido API
  version: 0.1.0
  description: |
    HTTP API contract. Default Docker Compose base URL uses host port 30880 (see infra/docker/.env.example).
    Implementation status and backlog are tracked in TASKS.md and PLAN.md; not all security controls are implemented yet.

    **Authentication:** When `SENTINEL_REQUIRE_AUTH=true` and `SENTINEL_JWT_SECRET` are set, all `/api/*` routes require `Authorization: Bearer <JWT>` (HS256).
    Claims: `sub` (string), `exp` (number), `role` (`analyst` or `admin`). When `SENTINEL_REQUIRE_AUTH` is unset or false, requests without a bearer token are accepted as an anonymous analyst (development only).
servers:
  - url: http://localhost:30880
    description: Docker Compose default (see infra/docker/.env.example)
  - url: http://localhost:8080
    description: Local uvicorn without Docker
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /ready:
    get:
      summary: Readiness check (database connectivity when DATABASE_URL is set)
      responses:
        '200':
          description: Ready or running without database
        '503':
          description: Database configured but unreachable
  /api/version:
    get:
      summary: API version and service name
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Missing or invalid bearer token when auth is required
        '403':
          description: Insufficient role
  /api/projects:
    post:
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created
  /api/projects/{project_id}:
    get:
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /api/sync/requests:
    post:
      summary: Ingest normalized requests from Caido bridge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestSyncPayload'
      responses:
        '202':
          description: Accepted
        '409':
          description: Scope manifest missing or invalid for the project
  /api/sync/findings:
    post:
      summary: Ingest findings from Caido bridge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindingSyncPayload'
      responses:
        '202':
          description: Accepted
        '409':
          description: Scope manifest missing or invalid for the project
  /api/projects/{project_id}/surface:
    get:
      summary: Get surface map
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /api/hypotheses/generate:
    post:
      summary: Generate bounded hypotheses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypothesisGenerationRequest'
      responses:
        '202':
          description: Accepted
        '409':
          description: Scope manifest missing or invalid for the project
  /api/hypotheses/{hypothesis_id}/approve:
    post:
      summary: Approve a hypothesis for execution as a bounded template
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: hypothesis_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Approved
        '409':
          description: Scope manifest missing or invalid for the project
  /api/feedback:
    post:
      summary: Record analyst feedback
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '201':
          description: Created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT with sub, exp, and role (analyst or admin)
  schemas:
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        owner_team:
          type: string
    RequestItem:
      type: object
      required: [caido_request_id, method, url, host, path]
      properties:
        caido_request_id:
          type: string
        method:
          type: string
        url:
          type: string
        host:
          type: string
        path:
          type: string
        status_code:
          type: integer
        req_headers_json:
          type: object
        resp_headers_json:
          type: object
    RequestSyncPayload:
      type: object
      required: [project_id, requests]
      properties:
        project_id:
          type: string
          format: uuid
        requests:
          type: array
          items:
            $ref: '#/components/schemas/RequestItem'
    FindingSyncPayload:
      type: object
      required: [project_id, findings]
      properties:
        project_id:
          type: string
          format: uuid
        findings:
          type: array
          items:
            type: object
    HypothesisGenerationRequest:
      type: object
      required: [project_id]
      properties:
        project_id:
          type: string
          format: uuid
        max_results:
          type: integer
          default: 10
    FeedbackRequest:
      type: object
      required: [object_type, object_id, feedback_type, user_id]
      properties:
        object_type:
          type: string
        object_id:
          type: string
          format: uuid
        feedback_type:
          type: string
        user_id:
          type: string
        notes:
          type: string